linting linker.sh, overhaul bashrc for colors and some aliases, tweak mpsyt settings

+++ linker.sh
@@ -1,28 +1,28 @@
 #!/bin/sh
 # Symbolic links for all the config files
-DOTFILES=$HOME/dotfiles
-SCRIPTS=$DOTFILES/.config/scripts/*
+DOTFILES="$HOME/dotfiles"
+scripts="$DOTFILES/.config/scripts/*"
 
-for S in $SCRIPTS; do
-  chmod 744 $S
+for s in $scripts; do
+  chmod 744 "$s"
 done
 
-[ ! -d $HOME/.vim ] && mkdir $HOME/.vim
+[ ! -d "$HOME/.vim" ] && mkdir "$HOME/.vim"
 
-LIST="
+list="
   .vim/custom
-  .Xresources .tmux.conf .xinitrc .bash_profile .gtkrc-2.0 .streamlinkrc
-  .vim/vimrc
+  .Xresources .tmux.conf .xinitrc .bash_profile .bashrc
+  .vim/vimrc .gtkrc-2.0 .streamlinkrc
 
   .config/i3 .config/scripts .config/gtk-3.0
   .config/wallpaper.jpg .config/mps-youtube/config
 "
-for target in $LIST; do
+for target in $list; do
   echo "$target"
   directory=${target%/*}
   # Make the parent directory if it does not exist
-  [ "$target" != "$directory" ] && [ ! -d $HOME/$directory ] \
-    && mkdir -p $HOME/$directory
-  [ -n "$target" ] && rm -fr "$HOME/$target"
+  [ "$target" != "$directory" ] && [ ! -d "$HOME/$directory" ] \
+    && mkdir -p "$HOME/$directory"
+  [ -n "$target" ] && rm -fr "${HOME:?}/$target"
   ln -s "$DOTFILES/$target" "$HOME/$target"
 done
